backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
langchain-groq
python-dotenv
pytest